some litle cleanups
