+++ OctoAwesome/OctoAwesome.Basics/OctoAwesome.Basics.csproj
@@ -73,23 +73,13 @@
     <Compile Include="CollisionPlane.cs" />
     <Compile Include="Definitions\Items\WauziEggDefinition.cs" />
     <Compile Include="Entities\WauziEntity.cs" />
-    <Compile Include="EntityComponents\BodyComponent.cs" />
-    <Compile Include="EntityComponents\ControllableComponent.cs" />
-    <Compile Include="EntityComponents\HeadComponent.cs" />
     <Compile Include="EntityComponents\InventoryComponent.cs" />
-    <Compile Include="EntityComponents\PositionComponent.cs" />
-    <Compile Include="EntityComponents\RenderComponent.cs" />
     <Compile Include="EntityComponents\ToolBarComponent.cs" />
     <Compile Include="Extension.cs" />
     <Compile Include="Biomes\FlatlandBiome.cs" />
     <Compile Include="Biomes\HighMountainBiome.cs" />
     <Compile Include="Biomes\HillsBiome.cs" />
     <Compile Include="Biomes\IBiome.cs" />
-    <Compile Include="EntityComponents\BodyPowerComponent.cs" />
-    <Compile Include="EntityComponents\CollisionComponent.cs" />
-    <Compile Include="EntityComponents\MassComponent.cs" />
-    <Compile Include="EntityComponents\BoxCollisionComponent.cs" />
-    <Compile Include="EntityComponents\EntityCollisionComponent.cs" />
     <Compile Include="Definitions\Blocks\BirchWoodBlockDefinition.cs" />
     <Compile Include="Definitions\Blocks\CactusBlockDefinition.cs" />
     <Compile Include="Definitions\Blocks\GravelBlockDefinition.cs" />
@@ -100,10 +90,6 @@
     <Compile Include="Definitions\Blocks\PlankBlockDefinition.cs" />
     <Compile Include="Definitions\Trees\BirchTreeDefinition.cs" />
     <Compile Include="Definitions\Trees\CactusTreeDefinition.cs" />
-    <Compile Include="EntityComponents\ForceComponent.cs" />
-    <Compile Include="EntityComponents\GravityComponent.cs" />
-    <Compile Include="EntityComponents\MoveableComponent.cs" />
-    <Compile Include="EntityComponents\PowerComponent.cs" />
     <Compile Include="ITreeDefinition.cs" />
     <Compile Include="Noise\INoise.cs" />
     <Compile Include="Biomes\LandBiomeGenerator.cs" />
@@ -145,15 +131,7 @@
     <Compile Include="Definitions\Blocks\WaterBlockDefinition.cs" />
     <Compile Include="Definitions\Blocks\WoodBlockDefinition.cs" />
     <Compile Include="Definitions\Trees\SpruceTreeDefinition.cs" />
-    <Compile Include="SimComponents\PhysicEngineComponent.cs" />
-    <Compile Include="SimulationComponents\BlockInteractionComponent.cs" />
-    <Compile Include="SimulationComponents\CollisionComponent.cs" />
-    <Compile Include="SimulationComponents\ForceAggregatorComponent.cs" />
-    <Compile Include="SimulationComponents\MoveComponent.cs" />
-    <Compile Include="SimulationComponents\AccelerationComponent.cs" />
-    <Compile Include="SimulationComponents\NewtonGravitatorComponent.cs" />
-    <Compile Include="SimulationComponents\PowerAggregatorComponent.cs" />
-    <Compile Include="SimulationComponents\WattMoverComponent.cs" />
+    <Compile Include="SimulationComponents\PhysicEngineComponent.cs" />
     <Compile Include="TreeDefinition.cs" />
     <Compile Include="TreePopulator.cs" />
     <Compile Include="Definitions\Items\PickAxeDefinition.cs" />
