+++ OctoAwesome/OctoAwesome/Entities/IControllable.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace OctoAwesome.Entities
+{
+    public interface IControllable
+    {
+        Coordinate Position { get; }
+        IController Controller { get; }
+        void Register(IController controller);
+        void Reset();
+    }
+}
